backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai>=1.0.0
requests>=2.32.3
aiohttp>=3.9.0
feedparser
beautifulsoup4
vaderSentiment
thefuzz>=0.19.0
apscheduler
sqlalchemy
bcrypt
pyJWT